content-security-policy/to-port-from-blink/inline-style-attribute-blocked-expected.txt

CONSOLE ERROR: line 10: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'". Either the 'unsafe-inline' keyword, a hash ('sha256-LpmHRjuhMOJl5UQVWi3ONaveclDEAfUVZg-1-9VZ57c='), or a nonce ('nonce-...') is required to enable inline execution.

PASS
